allow DataLayerServer to share the same DataLayerServerService
